requirements.txt
selenium == 3.141.0
pytest == 5.3.1
allure-pytest == 2.8.6
